+++ src/com/landawn/abacus/util/MutableFloat.java
@@ -84,10 +84,8 @@ public final class MutableFloat extends Number implements Comparable<MutableFloa
      * 
      * @param value  the value to set
      */
-    public MutableFloat setValue(final float value) {
+    public void setValue(final float value) {
         this.value = value;
-
-        return this;
     }
 
     public float getAndSet(final float value) {